scripts/requirements.txt
GitPython
requests
boto
